footer section moved at shared folder and recomended spelling updated
