+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router";
 import Navbar from "./components/shared/Navbar";
-import Footer from "./components/Home/Footer";
+import Footer from "./components/shared/Footer";
 
 export default function App() {
   return (
